refac(shared): rename exception package
