+++ shared/src/main/java/me/saiintbrisson/minecraft/command/exceptions/InsufficientPermissionsException.java
@@ -1,4 +1,4 @@
-package me.saiintbrisson.minecraft.command.exception;
+package me.saiintbrisson.minecraft.command.exceptions;
 
 import lombok.Getter;
